Move editor page into its own folder and add floating extended button Add Node
